Add readers and writers for function <-> lambda conversions
