+++ CHANGELOG.md
@@ -3,15 +3,15 @@
 ### Highlights :tada:
 + The JVM interface to Python has been completely rewritten from scratch to share all of its logic with the Scala Native backend by binding directly to CPython with JNA. This means that moving forward, ScalaPy JVM and Native will always have the same features and use near-identical logic for talking to Python libraries
 + Readers and Writers have been simplified to always work in terms of Python interpreter values, simplifying the implementation and reducing intermediate allocations
-+ Adds support for sending Scala functions into Python as lambdas
-+ Adds proxy collections to avoid copying data
++ Adds support for sending Scala functions into Python as lambdas, and reading Python lambdas into Scala functions
++ Adds proxy collections, which are Python objects that wrap over Scala collections instead of copying the underlying data
 + Introduces `StaticModule` API for loading modules into statically-typed objects
 + Expands support of `bracketAccess` to any `py.Any` type for the index
 + Adds `bracketUpdate` API to perform the equivalent of a Python `obj[key] = newValue`
 
 ### Breaking Changes :warning:
 + The `CPythonInterpreter` object should now be used to access low-level interpreter functions instead of `py.interpreter`
-+ Drops automatic conversion of Scala collections to Python in favor of `toPythonCopy` and `toPythonProxy` methods
++ Drops automatic conversion of Scala collections to Python in favor of `toPythonCopy` and `toPythonProxy` methods from the `py.SeqConverters` extension
 + Renames `arrayAccess` to `bracketAccess`
 
 ### Bug Fixes :bug:
