+++ core/shared/src/main/scala/me/shadaj/scalapy/py/package.scala
@@ -45,6 +45,16 @@ package object py {
     }
   }
 
+  implicit class SeqConverters[T, C <% Seq[T]](seq: C) {
+    def toPythonCopy(implicit elemWriter: Writer[T]): Any = {
+      Any.populateWith(CPythonInterpreter.createListCopy(seq, elemWriter.write))
+    }
+
+    def toPythonProxy(implicit elemWriter: Writer[T]): Any = {
+      Any.populateWith(CPythonInterpreter.createListProxy(seq, elemWriter.write))
+    }
+  }
+
   import py.{Dynamic => PyDynamic, Any => PyAny}
   trait PyQuotable {
     def stringToInsert: String
